rename sealmigration test package
