+++ vault/external_tests/sealmigration/seal_migration_pre14_test.go
@@ -1,6 +1,6 @@
 // +build !enterprise
 
-package seal_migration
+package sealmigration
 
 import (
 	"context"
